# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EstoqueClinic API",
    "version": "1.0.0",
    "description": "API REST para gestão de estoque de clínicas estéticas com multi-tenancy",
    "contact": {
      "name": "EstoqueClinic Support",
      "email": "suporte@estoqueclinic.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento Local"
    },
    {
      "url": "https://api.estoqueclinic.com",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de autenticação. Header: Authorization: Bearer {token}"
      }
    },
    "schemas": {
...
      "Clinic": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "clinic_group_id": {"type": "integer"},
          "name": {"type": "string"},
          "cnpj": {"type": "string"},
          "cro_number": {"type": "string"},
          "address_street": {"type": "string"},
          "address_number": {"type": "string"},
          "address_complement": {"type": "string"},
          "address_neighborhood": {"type": "string"},
          "address_city": {"type": "string"},
          "address_state": {"type": "string"},
          "address_zipcode": {"type": "string"},
          "phone": {"type": "string"},
          "email": {"type": "string", "format": "email"},
          "website": {"type": "string"},
          "logo_url": {"type": "string"},
          "timezone": {"type": "string"},
          "settings": {"type": "object"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/clinics": {
      "get": {
        "tags": ["Clinics"],
        "summary": "Listar clínicas do grupo",
        "description": "Lista todas as clínicas do grupo do usuário",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}},
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "is_active", "in": "query", "schema": {"type": "boolean"}}
        ],
        "responses": {
          "200": {
            "description": "Lista de clínicas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clinics": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Clinic"}
                    },
                    "pagination": {"type": "object"}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Clinics"],
        "summary": "Criar clínica",
        "description": "Cria uma nova clínica no grupo do usuário",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string"},
                  "cnpj": {"type": "string"},
                  "cro_number": {"type": "string"},
                  "address_street": {"type": "string"},
                  "address_number": {"type": "string"},
                  "address_complement": {"type": "string"},
                  "address_neighborhood": {"type": "string"},
                  "address_city": {"type": "string"},
                  "address_state": {"type": "string"},
                  "address_zipcode": {"type": "string"},
                  "phone": {"type": "string"},
                  "email": {"type": "string", "format": "email"},
                  "website": {"type": "string"},
                  "logo_url": {"type": "string"},
                  "timezone": {"type": "string"},
                  "settings": {"type": "object"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Clínica criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "clinic": {"$ref": "#/components/schemas/Clinic"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}